excmds: add "quit" and "q" commands
Closing the final tab via "tabclose" doesn't actually close
the window. Instead, we just get an empty placeholder tab.
Let's instead add a "quit" command which can be used to quit
the current tab, and if it's the last tab in the window,
close the window.

This is similar to how vim (and pentadactyl) work, which
lets you ":q" your way through any open buffers.