Support bounded tabmove even for absolute moves.
This actually simplifies the implementation since
we can handle both relative and absolute moves
in the same way and only need to find the next
best candidate if the initial choice fails.

Following Colin's suggestion, we find this
candidate by querying only tabs with the same
pinned property, sort them in ascending
index order and take the first or last tab
if the initial choice was smaller than or greater
than the active index, respectively. This has
the effect of moving the tab in the desired
direction as  as possible.