Revert "src: Simplify HintState constructor"
This reverts commit 092e501db41d170099b8b725fbe4014c338f7722.

This was causing a bug identified by @glacambre and @Lexi where
sometimes the html function isn't defined when returning to a page with
history back.

Personally, I think it's probably webpack or Firefox's fault: content.ts
clearly imports html-tagged-template as it's first action.

If it's not their fault then it's html-tagged-template's fault: it must
be refusing to set window.html.