Implement tridactyl-viewsource
The problem with the current :viewsource is that it loads a page in
which Tridactyl is not loaded. This commit fixes this by enabling a new
setting, "viewsource" which can be set to either "default" or
"tridactyl". If "tridactyl" is selected, then :viewsource will load the
source code of the HTML page in the current page instead of redirecting
the user to the view-source:// version of the page.