excmds.ts: make scrollto() call recursiveScroll when failing to scroll
Scrollto can fail to scroll on some pages such as the ones created by
PDF.js. When this happens, and if the user is asking to go to the very
top or very bottom of the page, try using recursiveScroll instead.

It's tempting to skip the window.scrollTo call and just use
recursiveScroll no matter what but recursiveScroll doesn't always work
either. For example, window.scrollTo can take the user to the bottom of
the page on mozilla's doc while recursiveScroll cannot.

This fixes https://github.com/cmcaine/tridactyl/issues/382