Merge input and insert mode
Merging input and insert mode is done by removing every occurence of the
input mode. Instead, where Tridactyl would enter input mode, it instead
now sets a "focusinput" boolean to true in Tridactyl's state. This
boolean is set to false every time the mode changes, no matter what mode
Tridactyl was previously in.

Having a separate boolean instead of a mode requires separate handling
in both keydown_content.ts (for key suppression) and
parsers/insertmode.ts (for actually calling `focusinput` when Tab is
pressed).

This fixes half of https://github.com/cmcaine/tridactyl/issues/455 .
The bug happened because modeSpecificSuppression() was only called when
state.mode was "input" and the cursor wasn't in a isTextEditable
element. It is fixed by making the call to modeSpecificSuppression
unconditional.