Use fillcmdline to provide feeback from yank [Fixes #138]
We may want something better once we get a proper UI, but for now, we
can tell the user that yank worked and let them know exactly what was
yanked.  The output follows the convention of the instructions for
installing native messenging: we have a comment in the command line so
nothing happens except the command line closes when the user presses
enter.