content.{ts,css}: Make modeindicator disappear on hover
This fixes https://github.com/cmcaine/tridactyl/issues/565
I chose to append a TridactylInvisible class instead of setting the
modeindicator setting to false or manually removing the element from the
DOM because this seemed more robust (e.g. when "modeindicator" is set to
true, you expect it to stay set to true and for the element to exist
within the DOM).