{excmds,controller}.ts: Fix race condition in state.mode synchronization
This fixes https://github.com/cmcaine/tridactyl/issues/613.
This was a really fun bug. What happened was this:
- First, the content script set state.mode to "hint", which was then
  synchronized
- The event listener in the background script noticed the update and set
  state.mode to "hint" in the background script
- Then, the content script translated the hint selection into an excmd
  that needed to be executed in the background script and sent said
  command to the controller, which lives in the background script
- The content script then set state.mode to "normal"
- The background script executed the command and saved it in
  state.last_ex_str, which was then synchronized
- The event listener in the content script noticed the update and set
  last_ex_str to the last executed str and "state.mode" to "hint"

So basically, the problem was that the background script didn't notice
the state.mode update right after it happened. I fixed this by moving
last_ex_str out of "state" since it doesn't need to be synchronized with
the content script.

This means that the same kind of race condition can still happen. I'm
not sure how to fix this. We could just kill state completely and
instead have state be updated through message passing, but this probably
wouldn't be very ergonomic.
Another solution, the one envisioned for Tridactylv2, is to move to the
content script entirely. This is probably the best option.