excmds: use bullet list for "related settings"
When these comments get formatted into ":help open", etc,
the line breaks are eaten and we get a big jumble of text.
Let's space things out so that they get formatted in a sane
way.

Possibly a definition list would be better than bullets, but
this is at least an improvement over the status quo.