{excmds,webext,messaging}.ts: Remove l() promise logger
When a promise throws an error in l(), the error is automatically logged
in the console. This is a problem because it happens even when the error
is catched by the code that called l(), which might not want to log
errors at all.

l() was necessary when errors thrown in promises didn't reach the
toplevel. Now that they do, it is safe to remove l().

This will enable better error handling.