Fix scrolling on bugzilla.mozilla.org
The problem was caused by elements having their scroll-behavior set to
"smooth". This meant that offsetTop was updated with a delay. The delay
made Tridactyl that the element couldn't be scrolled at all and thus
scrolling failed.

This commit has the nice side-effect of disabling smooth scrolling ig
"smoothscroll" isn't set to "true" in the config.

This fixes https://github.com/cmcaine/tridactyl/issues/762.