native_background.ts: Fix clipboard() adding newlines to copied strings
Here-documents require a newline at the end of their content and there's
no way around it. In order to prevent the newline to be sent to the
clipboard command, we can do two things: not use here-documents or pipe
the content to sed first to remove the newline.

Not using here-documents means doing a lot of string manipulations to
escape the string properly. Using here-documents is conceptually
simpler.

Hence, we keep using here-documents and pipe the string to sed in order
to remove the newline.