Fix Tridacyl not working on about:blank newtab page.
Apparently, Firefox can't give you key events for the window if <body>
is `display: none`. This is fixed by turning the page transparent
instead of not displaying it. Since the scrollbar is still visible when
the page is transparent, we also set the page's height to 0px.

All of these modifications are reversed when the page needs to be
displayed.