History.ts: Add missing space when inserting completion options
e5d2be7 introduced a bug while trying to fix
https://github.com/cmcaine/tridactyl/issues/838: when the command line
contains `tabopen -flag ` and the user tries to insert the selected
completion option in the command line by pressing <Space>, the resulting
command line will be `tabopen -flagurl` instead of `tabopen -flag url`.

This is fixed by appending a trailing space to the `options` variable if
it isn't an empty string.