Fix bufferall selecting the wrong window.
The problem was that when calling sort on an array of numbers without
specifying the sort function that should be used, the array will be
sorted according to the lexicographic order of the elements.

This is fixed by specifying what function should be used to sort the
array.

This fixes https://github.com/cmcaine/tridactyl/issues/819.