Refactor hinting.pipe to return a tuple.
Now returns a tuple of the element and the number of hints available
when the function was run. This mainly fixes the rapid hint bug
encountered when the user enters rapid hint mode with only one link
available.

Another application idea would be modeline info or something in that
vein.