Fix JavaScript-Python Unicode issues on Windows
The following patches fix Unicode handling issues between JavaScript
and Python on Windows.

The patches introduce encoding of native-messages (with URI and
Base64 encoding) exchanged between JavaScript and Python.

In short, we exchange Base64 "over the wire" to avoid Unicode
characters causing interoperability issues, and URI encoding is
necessary for `atob()/btoa()` to be successful with UTF-8 as
discussed here [0].

This patch requires lower-level changes around native-messaging both
on the JavaScript and Python end. So the users must upgrade to
native messenger version `0.1.9` _and_ the Tridactyl add-on before
using this.

Thank you.

[0] https://developer.mozilla.org/en-US/docs/Web/API/WindowBase64/Base64_encoding_and_decoding#The_Unicode_Problem