background.ts: Ignore errors when TabLeft/TabEnter can't be triggered
When switching to or from a tab, Tridactyl sends a message to the
previously and newly-active tabs in order to trigger the TabLeft and
TabEnter autocommands.

This didn't work with privileged tabs as these tabs do not have a
content script. Instead, an exception was thrown. This commit adds
exception handling for these exceptions.