config.ts: Turn default_config into a class
This commit turns default_config into a class. This is useful because
TypeDoc's layout for classes is better than the one for object literals.

On top of that, default_config being a class will help with adding type
information to settings, which will in turn let us generate config
metadata that we will then be able to use to typecheck `:set` and to
provide more information in completions.