Implement ownTab{,Id,Container}, use them to set modeindicator color
Firefox used to give the same color to all modeindicators on startup
with browser.sessionstore.restore_on_demand set to true in about:config,
This was caused by the use of `activeTabContainer` when setting the
modeindicator in content.ts. This call is replaced by a call to
`ownTabContainer`, which fixes this bug.

Closes https://github.com/tridactyl/tridactyl/issues/816 .