Implement mergedeep function
Previously, running `config.get()` would result in only top level
objects being merged and returned. This means that if you ran
`config.get()`, the `nmaps` object would be the default one if no custom
bindings were defined, but it would not contain the default bindings if
custom bindings were defined.

This commit fixes that by implementing a mergeDeep function that makes
sure that children of the returned are merged with the default config.

Closes https://github.com/tridactyl/tridactyl/issues/390.