Get Buffer and BufferAll code a bit closer together
Including implementing # for BufferAll... sort of. The completion
options _sort_ right, but `:BufferAll #` doesn't work and I can't
quite be bothered to figure out why. Probably something to do with not
copying the Fuse scorer from Buffer to BufferAll?

This was mostly a short refactor to make sure I understood what was
going on with the buffer completion sources, which are what I'm using
as the test case for a new completions architecture.