Turn `:bmarks` into a proper excmd
Turning `:bmarks` into a proper excmd has several advantages: it will
show up in the `:help` pages, be available for excmd completions and
enbales adding specific arguments (such as "-t") in order to open
bookmarks in other tabs.

This time I'm directly messaging the background script in order to
trigger a tabopen rather than trying to bring tabopen to the content
script.

Closes https://github.com/tridactyl/tridactyl/issues/902.