Various tutorial improvements
This commit renames tutorial pages with the number in which they are
visited by the user. This helps when you need to add a new page and want
to link to the previous/next one in it.

It also adds a new page about the native messenger which describes what
it does and how to install it. I took this opportunity to rename
installnative to nativeinstall and to add an ex alias that goes the
other way.

Last, this commit  mentions that `:help` also accepts settings and keys
as argument and that we now have a troobleshooting guide.