config.ts: improve addChangeListener handling of objects
This commit prevents triggering of event listeners when values that
aren't set in USERCONFIG are set to the value they have in
default_config.

Also, before this commit, listeners added to setting objects (e.g.
`nmaps`) with addChangeListener would always be triggered, even if the
new object was structurally equivalent to the previous one. This commit
stringifies both object in order to perform comparisons, this prevents
this case from happening.

Note that there's still a problem with the way objects are handled: when
setting the key of an object (e.g. `nmaps.j`), eventListeners set on the
object (`nmaps`) will be triggered even if no user-facing change happens
(e.g. the user set `nmaps.j` to whatever value `nmaps.j` already had in
default_config). This is because the user config is actually the result
of a merge between the default config and their custom config, but the
event listener code performs no such merges for simplicity's sake.